+++ apps/backoffice-v2/src/common/components/atoms/RouteError/RouteErrorWithProviders.tsx
@@ -1,0 +1,10 @@
+import { Providers } from '@/common/components/templates/Providers/Providers';
+import { RouteError } from '@/common/components/atoms/RouteError/RouteError';
+
+export const RouteErrorWithProviders = () => {
+  return (
+    <Providers>
+      <RouteError />
+    </Providers>
+  );
+};