+++ apps/backoffice-v2/src/main.tsx
@@ -7,6 +7,8 @@ import '@ballerine/ui/dist/style.css';
 import '@fontsource/inter';
 
 import { Toaster } from '@/common/components/organisms/Toaster/Toaster';
+import { registerFont } from '@ballerine/react-pdf-toolkit';
+import { Font } from '@react-pdf/renderer';
 import { Router } from './Router/Router';
 import { env } from './common/env/env';
 import './i18n';
@@ -15,6 +17,8 @@ import dayjs from 'dayjs';
 import advancedFormat from 'dayjs/plugin/advancedFormat';
 
 dayjs.extend(advancedFormat);
+
+registerFont(Font);
 
 export const TOAST_DURATION_IN_MS = 1000 * 3;