+++ apps/backoffice-v2/src/pages/Entity/components/Case/Case.Actions.tsx
@@ -3,6 +3,7 @@ import { Badge } from '@ballerine/ui';
 import { FunctionComponent } from 'react';
 
 import { ActionsVariant } from '@/pages/Entity/components/Case/actions-variants/ActionsVariant/ActionsVariant';
+import { CaseOptions } from '@/pages/Entity/components/Case/components/CaseOptions/CaseOptions';
 import { AssignDropdown } from '../../../../common/components/atoms/AssignDropdown/AssignDropdown';
 import { ctw } from '../../../../common/utils/ctw/ctw';
 import { tagToBadgeData } from './consts';
@@ -40,7 +41,7 @@ export const Actions: FunctionComponent<IActionsProps> = ({
 
   return (
     <div className={`col-span-2 space-y-2 bg-base-100 px-4 pt-4`}>
-      <div className={`mb-8 flex flex-row space-x-3.5`}>
+      <div className={`mb-8 flex flex-row justify-between space-x-3.5`}>
         <AssignDropdown
           assignedUser={assignedUser}
           assignees={assignees}
@@ -50,8 +51,9 @@ export const Actions: FunctionComponent<IActionsProps> = ({
           authenticatedUserId={authenticatedUser?.id}
           isDisabled={isWorkflowCompleted}
         />
+        <CaseOptions />
       </div>
-      <div className={`flex h-20 justify-between gap-4`}>
+      <div className={`min-h-20 flex justify-between gap-4`}>
         <div className={`flex flex-col space-y-3`}>
           <h2
             className={ctw(`w-full max-w-[35ch] break-all text-4xl font-semibold leading-9`, {