+++ apps/backoffice-v2/src/pages/Entity/pdfs/case-information/components/CaseInformationPageContainer/CaseInformationPageContainer.tsx
@@ -1,0 +1,9 @@
+import { FunctionComponentWithChildren } from '@/common/types';
+import { tw } from '@ballerine/react-pdf-toolkit';
+import { Page, View } from '@react-pdf/renderer';
+
+export const CaseInformationPageContainer: FunctionComponentWithChildren = ({ children }) => (
+  <Page wrap={false}>
+    <View style={tw('flex flex-col p-5')}>{children}</View>
+  </Page>
+);