+++ apps/backoffice-v2/src/pages/Workflows/Workflows.page.tsx
@@ -1,0 +1,5 @@
+import React, { FunctionComponent } from 'react';
+
+export const Workflows: FunctionComponent = () => {
+  return <div>Workflows.page</div>;
+};