+++ packages/eslint-config-react/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @ballerine/eslint-config-react
 
+## 2.0.4
+
+### Patch Changes
+
+- Bump
+- Updated dependencies
+  - @ballerine/eslint-config@1.1.4
+
+## 2.0.3
+
+### Patch Changes
+
+- document changes
+- Updated dependencies
+  - @ballerine/eslint-config@1.1.3
+
 ## 2.0.2
 
 ### Patch Changes