+++ sdks/workflow-node-sdk/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @ballerine/workflow-node-sdk
 
+## 0.6.17
+
+### Patch Changes
+
+- Bump
+- Updated dependencies
+  - @ballerine/workflow-core@0.6.17
+
+## 0.6.16
+
+### Patch Changes
+
+- document changes
+- Updated dependencies
+  - @ballerine/workflow-core@0.6.16
+
+## 0.6.15
+
+### Patch Changes
+
+- @ballerine/workflow-core@0.6.15
+
 ## 0.6.14
 
 ### Patch Changes