+++ services/websocket-service/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @ballerine/websocket-service
 
+## 0.1.4
+
+### Patch Changes
+
+- Bump
+
+## 0.1.3
+
+### Patch Changes
+
+- document changes
+
 ## 0.1.2
 
 ### Patch Changes